% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-bucket.R
\name{create_bucket}
\alias{create_bucket}
\title{Create a new S3 bucket}
\usage{
create_bucket(bucket, acl = c("bucket-owner-full-control", "private",
  "public-read", "public-read-write", "aws-exec-read", "authenticated-read",
  "bucket-owner-read"), location_constraint = NULL, grant_full_control = "",
  grant_read = "", grant_read_acp = "", grant_write = "",
  grant_write_acp = "", aws_access_key_id = NULL,
  aws_secret_access_key = NULL, aws_session_token = NULL,
  region_name = NULL, profile_name = NULL)
}
\arguments{
\item{bucket}{name of the bucket. Not every string is an acceptable bucket name.
For information on bucket naming restrictions, see
\href{http://docs.aws.amazon.com/AmazonS3/latest/dev/UsingBucket.html}{Working with Amazon S3 Buckets}.}

\item{acl}{the canned ACL to apply to the bucket. Options are "\code{private}",
"\code{public-read}", "\code{public-read-write}", "\code{aws-exec-read}", "\code{authenticated-read}",
"\code{bucket-owner-read}", "\code{bucket-owner-full-control}". The default is
"\code{bucket-owner-full-control}".}

\item{location_constraint}{Specifies the region where the bucket will be created.
If you don't specify a region, the bucket will be created in US Standard.}

\item{grant_full_control, grant_read, grant_read_acp, grant_write, grant_write_acp}{grantee strings (optional)}

\item{aws_access_key_id}{AWS access key id}

\item{aws_secret_access_key}{AWS secret access key}

\item{aws_session_token}{AWS session token}

\item{profile_name}{profile name}
}
\value{
the API call result structure (invisibly). The \code{Location} element will
have the bucket name if the call was successful.
}
\description{
To create a bucket, you must register with Amazon S3 and have a valid AWS Access
Key ID to authenticate requests. Anonymous requests are never allowed to create
buckets. By creating the bucket, you become the bucket owner.
}
\details{
By default, the bucket is created in the US East (N. Virginia) region. You can
optionally specify a region in the request body.
}
\examples{
\dontrun{
}
}
\references{
\url{https://boto3.readthedocs.io/en/latest/reference/services/s3.html#S3.Client.create_bucket}
}
