% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete-bucket.R
\name{delete_bucket}
\alias{delete_bucket}
\title{Delete an S3 bucket}
\usage{
delete_bucket(bucket, aws_access_key_id = NULL,
  aws_secret_access_key = NULL, aws_session_token = NULL,
  region_name = NULL, profile_name = NULL)
}
\arguments{
\item{bucket}{name of the bucket}

\item{aws_access_key_id}{AWS access key id}

\item{aws_secret_access_key}{AWS secret access key}

\item{aws_session_token}{AWS session token}

\item{profile_name}{profile name}
}
\value{
the API call result structure (invisibly). The \code{Location} element will
have the bucket name if the call was successful.
}
\description{
Deletes the bucket. All objects (including all object versions and Delete Markers)
in the bucket must be deleted before the bucket itself can be deleted.
}
\examples{
\dontrun{
}
}
\references{
\url{https://boto3.readthedocs.io/en/latest/reference/services/s3.html#S3.Client.create_bucket}
}
