% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file-exists.R
\name{file_exists}
\alias{file_exists}
\title{Test if a file exists in the specified bucket}
\usage{
file_exists(bucket, key, aws_access_key_id = NULL,
  aws_secret_access_key = NULL, aws_session_token = NULL,
  region_name = NULL, profile_name = NULL)
}
\arguments{
\item{bucket}{name of the bucket to download from.}

\item{key}{name of the key (file).}

\item{aws_access_key_id}{AWS access key id}

\item{aws_secret_access_key}{AWS secret access key}

\item{aws_session_token}{AWS session token}

\item{region_name}{region name}

\item{profile_name}{profile name}
}
\description{
Test if a file exists in the specified bucket
}
\examples{
\dontrun{
file_exists("mybucket", "hello.txt")
}
}
\references{
\url{https://boto3.readthedocs.io/en/latest/reference/services/s3.html#S3.Client.download_file}
}
