% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roto.s3-package.R
\docType{package}
\name{roto.s3}
\alias{roto.s3}
\alias{roto.s3-package}
\title{Access and Orchestrate Amazon Simple Storage Service..}
\description{
This is a 'reticulated' wrapper for the 'Python' 'boto3' 'AWS'
'S3' client library \url{https://boto3.readthedocs.io/en/latest/reference/services/s3.html}.
It requires 'Python' version 3.5+ and an 'AWS' account.
}
\note{
This package \strong{requires} Python >= 3.5 to be available and the
\code{boto3} Python module. The package author highly recommends setting
\code{RETICULATE_PYTHON=/usr/local/bin/python3} in your \code{~/.Renviron} file
to ensure R + \code{reticulate} will use the proper python version.
}
\author{
Bob Rudis (bob@rud.is)
}
